Fix warnings. Rename file to remove conflict with parser on windows
