+++ requirements.txt
@@ -1,3 +1,3 @@
-Jinja2==2.10.1
+Jinja2==2.11.2
 requests==2.24.0
 psutil==5.7.3
