+++ src/cbc_syslog/resource_fetcher.py
